requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
boto3==1.4.7
botocore==1.7.7           # via boto3, s3transfer
configparser==3.5.0       # via flake8
docutils==0.14
enum34==1.1.6             # via flake8
flake8==3.4.1
futures==3.1.1            # via s3transfer
isort==4.2.15
jmespath==0.9.3           # via boto3, botocore
mccabe==0.6.1             # via flake8
multilint==2.1.0
py==1.4.34                # via pytest
pycodestyle==2.3.1        # via flake8
pyflakes==1.5.0           # via flake8
pygments==2.2.0
pytest==3.2.2
python-dateutil==2.6.1    # via botocore
pyyaml==3.12
s3transfer==0.1.11        # via boto3
six==1.10.0
